Afterpay payment method <label> not clickable
Found an issue where the Afterpay payment method's <label> element is not clickable to select the methods radio button and set the payment method as _active. This is due to the id="afterpay" being added "payment-method" DIV element instead of the <input type="radio"> element. For form elements to be selectable via the associated <label> element the ID of the <input> must match the labels "for" attribute for example <label for="afterpay">. This is quite a large usability issue as customers will more often than not click or touch the Afterpay logo within the label element to select the payment method rather than selecting the radio button itself.

I also note the styling in the afterpay-checkout.less file also relies on this parent ID for styling. This would have to be updated to use a class="afterpay" instead so the ID can be moved to the input.